Refactored validation event handler into controllers lib
