+++ src/controllers/index.js
@@ -10,11 +10,60 @@ const Router = require('koa-router')
 // Load the Clean Architecture Use Case libraries.
 const useCases = require('../use-cases')
 
+// Load the Clean Architecture Adapters library
+const Adapters = require('../adapters')
+const adapters = new Adapters()
+
 // Load the REST API Controllers.
 const PostEntry = require('./post-entry')
 const GetAll = require('./get-all')
 
-module.exports = function attachRESTControllers (app) {
+// Top-level function for this library.
+// Start the various Controllers and attach them to the app.
+async function attachControllers (app) {
+  // Attach the REST controllers to the Koa app.
+  attachRESTControllers(app)
+
+  // Start the P2WDB and attach the validation event handler/controller to
+  // the add-entry Use Case.
+  await attachValidationController()
+}
+
+// Start the P2WDB and its downstream depenencies (IPFS, ipfs-coord, OrbitDB).
+// Also attach the post-validation, peer-replication event handler (controller)
+// to the Add-Entry Use Case.
+async function attachValidationController () {
+  try {
+    // Start the P2WDB.
+    await adapters.p2wdb.start()
+
+    // Trigger the addPeerEntry() use-case after a replication-validation event.
+    adapters.p2wdb.orbit.validationEvent.on(
+      'ValidationSucceeded',
+      async function (data) {
+        try {
+          // console.log(
+          //   'ValidationSucceeded event triggering addPeerEntry() with this data: ',
+          //   data
+          // )
+
+          await useCases.addEntry.addPeerEntry(data)
+        } catch (err) {
+          console.error(
+            'Error trying to process peer data with addPeerEntry(): ',
+            err
+          )
+          // Do not throw an error. This is a top-level function.
+        }
+      }
+    )
+  } catch (err) {
+    console.error('Error in controllers/index.js/startP2wdb()')
+    throw err
+  }
+}
+
+function attachRESTControllers (app) {
   const baseUrl = '/temp'
   const router = new Router({ prefix: baseUrl })
 
@@ -28,7 +77,6 @@ module.exports = function attachRESTControllers (app) {
   const postEntry = new PostEntry({
     addEntry: useCases.addEntry
   })
-
   // curl -H "Content-Type: application/json" -X POST -d '{ "user": "test" }' localhost:5001/temp/write
   router.post('/write', async (ctx, next) => {
     try {
@@ -53,3 +101,5 @@ module.exports = function attachRESTControllers (app) {
   // Attach the Controller routes to the Koa app.
   app.use(router.routes()).use(router.allowedMethods())
 }
+
+module.exports = { attachControllers }
