+++ src/use-cases/index.js
@@ -19,37 +19,37 @@ const readEntry = new ReadEntry({ p2wdb: adapters.p2wdb })
 
 // Start all the non-REST controllers for the P2WDB. This includes IPFS,
 // ipfs-coord, JSON RPC over IPFS, and OrbitDB with the custom access controller.
-async function startP2wdb () {
-  try {
-    // Start the P2WDB.
-    await adapters.p2wdb.start()
-
-    // Trigger the addPeerEntry() use-case after a replication-validation event.
-    adapters.p2wdb.orbit.validationEvent.on(
-      'ValidationSucceeded',
-      async function (data) {
-        try {
-          // console.log(
-          //   'ValidationSucceeded event triggering addPeerEntry() with this data: ',
-          //   data
-          // )
-
-          await addEntry.addPeerEntry(data)
-        } catch (err) {
-          console.error(
-            'Error trying to process peer data with addPeerEntry(): ',
-            err
-          )
-          // Do not throw an error. This is a top-level function.
-        }
-      }
-    )
-  } catch (err) {
-    console.error('Error in controllers/index.js/startP2wdb()')
-    throw err
-  }
-}
-startP2wdb()
+// async function startP2wdb () {
+//   try {
+//     // Start the P2WDB.
+//     await adapters.p2wdb.start()
+//
+//     // Trigger the addPeerEntry() use-case after a replication-validation event.
+//     adapters.p2wdb.orbit.validationEvent.on(
+//       'ValidationSucceeded',
+//       async function (data) {
+//         try {
+//           // console.log(
+//           //   'ValidationSucceeded event triggering addPeerEntry() with this data: ',
+//           //   data
+//           // )
+//
+//           await addEntry.addPeerEntry(data)
+//         } catch (err) {
+//           console.error(
+//             'Error trying to process peer data with addPeerEntry(): ',
+//             err
+//           )
+//           // Do not throw an error. This is a top-level function.
+//         }
+//       }
+//     )
+//   } catch (err) {
+//     console.error('Error in controllers/index.js/startP2wdb()')
+//     throw err
+//   }
+// }
+// startP2wdb()
 
 // Export the instances of the use-cases.
 module.exports = {
